Изменение модели Signup
